+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed GetAzRecoveryServicesVaultSettingsFile cmdlet used for downloading vault credential file.
 
 * Fixed issue in Disable-AzRecoveryServicesBackupProtection cmdlet.
 * Fixed issue in Enable-AzRecoveryServicesBackupProtection cmdlet.